+++ x-pack/plugin/runtime-fields/src/test/java/org/elasticsearch/xpack/runtimefields/mapper/DateFieldScriptTests.java
@@ -11,6 +11,7 @@ import org.apache.lucene.index.DirectoryReader;
 import org.apache.lucene.index.RandomIndexWriter;
 import org.apache.lucene.store.Directory;
 import org.apache.lucene.util.BytesRef;
+import org.elasticsearch.common.time.DateFormatter;
 import org.elasticsearch.index.mapper.MapperService;
 import org.elasticsearch.script.ScriptContext;
 import org.elasticsearch.search.lookup.SearchLookup;
@@ -19,28 +20,30 @@ import java.io.IOException;
 import java.util.List;
 import java.util.Map;
 
+import static org.hamcrest.Matchers.equalTo;
 import static org.mockito.Mockito.mock;
 
-public class BooleanScriptFieldTests extends ScriptFieldTestCase<BooleanFieldScript.Factory> {
-    public static final BooleanFieldScript.Factory DUMMY = (fieldName, params, lookup) -> ctx -> new BooleanFieldScript(
+public class DateFieldScriptTests extends ScriptFieldTestCase<DateFieldScript.Factory> {
+    public static final DateFieldScript.Factory DUMMY = (fieldName, params, lookup, formatter) -> ctx -> new DateFieldScript(
         fieldName,
         params,
         lookup,
+        formatter,
         ctx
     ) {
         @Override
         public void execute() {
-            emit(false);
+            emit(1595431354874L);
         }
     };
 
     @Override
-    protected ScriptContext<BooleanFieldScript.Factory> context() {
-        return BooleanFieldScript.CONTEXT;
+    protected ScriptContext<DateFieldScript.Factory> context() {
+        return DateFieldScript.CONTEXT;
     }
 
     @Override
-    protected BooleanFieldScript.Factory dummyScript() {
+    protected DateFieldScript.Factory dummyScript() {
         return DUMMY;
     }
 
@@ -48,21 +51,25 @@ public class BooleanScriptFieldTests extends ScriptFieldTestCase<BooleanFieldScr
         try (Directory directory = newDirectory(); RandomIndexWriter iw = new RandomIndexWriter(random(), directory)) {
             iw.addDocument(List.of(new StoredField("_source", new BytesRef("{}"))));
             try (DirectoryReader reader = iw.getReader()) {
-                BooleanFieldScript script = new BooleanFieldScript(
+                DateFieldScript script = new DateFieldScript(
                     "test",
                     Map.of(),
                     new SearchLookup(mock(MapperService.class), (ft, lookup) -> null),
+                    DateFormatter.forPattern(randomDateFormatterPattern()).withLocale(randomLocale(random())),
                     reader.leaves().get(0)
                 ) {
                     @Override
                     public void execute() {
-                        for (int i = 0; i <= AbstractFieldScript.MAX_VALUES * 1000; i++) {
-                            emit(i % 2 == 0);
+                        for (int i = 0; i <= AbstractFieldScript.MAX_VALUES; i++) {
+                            emit(0);
                         }
                     }
                 };
-                // There isn't a limit to the number of values so this won't throw
-                script.execute();
+                Exception e = expectThrows(IllegalArgumentException.class, script::execute);
+                assertThat(
+                    e.getMessage(),
+                    equalTo("Runtime field [test] is emitting [101] values while the maximum number of values allowed is [100]")
+                );
             }
         }
     }
