+++ x-pack/plugin/runtime-fields/src/test/java/org/elasticsearch/xpack/runtimefields/mapper/StringFieldScriptTests.java
@@ -11,7 +11,6 @@ import org.apache.lucene.index.DirectoryReader;
 import org.apache.lucene.index.RandomIndexWriter;
 import org.apache.lucene.store.Directory;
 import org.apache.lucene.util.BytesRef;
-import org.elasticsearch.common.time.DateFormatter;
 import org.elasticsearch.index.mapper.MapperService;
 import org.elasticsearch.script.ScriptContext;
 import org.elasticsearch.search.lookup.SearchLookup;
@@ -23,27 +22,26 @@ import java.util.Map;
 import static org.hamcrest.Matchers.equalTo;
 import static org.mockito.Mockito.mock;
 
-public class DateScriptFieldTests extends ScriptFieldTestCase<DateFieldScript.Factory> {
-    public static final DateFieldScript.Factory DUMMY = (fieldName, params, lookup, formatter) -> ctx -> new DateFieldScript(
+public class StringFieldScriptTests extends ScriptFieldTestCase<StringFieldScript.Factory> {
+    public static final StringFieldScript.Factory DUMMY = (fieldName, params, lookup) -> ctx -> new StringFieldScript(
         fieldName,
         params,
         lookup,
-        formatter,
         ctx
     ) {
         @Override
         public void execute() {
-            emit(1595431354874L);
+            emit("foo");
         }
     };
 
     @Override
-    protected ScriptContext<DateFieldScript.Factory> context() {
-        return DateFieldScript.CONTEXT;
+    protected ScriptContext<StringFieldScript.Factory> context() {
+        return StringFieldScript.CONTEXT;
     }
 
     @Override
-    protected DateFieldScript.Factory dummyScript() {
+    protected StringFieldScript.Factory dummyScript() {
         return DUMMY;
     }
 
@@ -51,17 +49,16 @@ public class DateScriptFieldTests extends ScriptFieldTestCase<DateFieldScript.Fa
         try (Directory directory = newDirectory(); RandomIndexWriter iw = new RandomIndexWriter(random(), directory)) {
             iw.addDocument(List.of(new StoredField("_source", new BytesRef("{}"))));
             try (DirectoryReader reader = iw.getReader()) {
-                DateFieldScript script = new DateFieldScript(
+                StringFieldScript script = new StringFieldScript(
                     "test",
                     Map.of(),
                     new SearchLookup(mock(MapperService.class), (ft, lookup) -> null),
-                    DateFormatter.forPattern(randomDateFormatterPattern()).withLocale(randomLocale(random())),
                     reader.leaves().get(0)
                 ) {
                     @Override
                     public void execute() {
                         for (int i = 0; i <= AbstractFieldScript.MAX_VALUES; i++) {
-                            emit(0);
+                            emit("test");
                         }
                     }
                 };
@@ -69,6 +66,37 @@ public class DateScriptFieldTests extends ScriptFieldTestCase<DateFieldScript.Fa
                 assertThat(
                     e.getMessage(),
                     equalTo("Runtime field [test] is emitting [101] values while the maximum number of values allowed is [100]")
+                );
+            }
+        }
+    }
+
+    public void testTooManyChars() throws IOException {
+        try (Directory directory = newDirectory(); RandomIndexWriter iw = new RandomIndexWriter(random(), directory)) {
+            iw.addDocument(List.of(new StoredField("_source", new BytesRef("{}"))));
+            try (DirectoryReader reader = iw.getReader()) {
+                StringFieldScript script = new StringFieldScript(
+                    "test",
+                    Map.of(),
+                    new SearchLookup(mock(MapperService.class), (ft, lookup) -> null),
+                    reader.leaves().get(0)
+                ) {
+                    @Override
+                    public void execute() {
+                        StringBuilder big = new StringBuilder();
+                        while (big.length() < StringFieldScript.MAX_CHARS / 4) {
+                            big.append("test");
+                        }
+                        String bigString = big.toString();
+                        for (int i = 0; i <= 4; i++) {
+                            emit(bigString);
+                        }
+                    }
+                };
+                Exception e = expectThrows(IllegalArgumentException.class, script::execute);
+                assertThat(
+                    e.getMessage(),
+                    equalTo("Runtime field [test] is emitting [1310720] characters while the maximum number of values allowed is [1048576]")
                 );
             }
         }
